+++ src/wai/common/_InvalidStateError.py
@@ -1,0 +1,5 @@
+class InvalidStateError(Exception):
+    """
+    Exception indicating an error in the internal state of an object.
+    """
+    pass