+++ src/wai/common/__init__.py
@@ -1,4 +1,6 @@
 from ._functions import is_hashable
+from ._Interval import Interval
+from ._InvalidStateError import InvalidStateError
 from ._memoise import Memoiser, MemoiserFactory, MEMO_EXTENSION, MemoFile, InvalidMemoFileError, PicklableDict
 from ._pool import run_on_all, num_processes
 from ._switch import switch, case, default, break_